Pickup Event is now recognized
I was able to get confirm that the pickup event was being registered properly. I haven't yet figured out how to attach the ball to the player mesh. There is a bug where the event won't be recognized after a score has been made. I will fix this in my next set of changes in the next few days.